requirements.txt
typer>=0.9.0
rich>=13.0.0
prompt-toolkit>=3.0.0
python-dotenv>=1.0.0
python-docx>=0.8.11
